fix user can always change his rating

+++ movie_rating/views.py
@@ -28,9 +28,13 @@ def detail_movie(request, slug):
         form = RatingForm(request.POST)
         if form.is_valid():
             rating = request.POST.get('rating')
-            Rating.objects.filter(user=request.user, movie=movie).delete()
-            Rating.objects.create(user=request.user, rating=rating, movie=movie)
-            messages.success(request, 'Thank you for your review of the movie!')
+            if Rating.objects.filter(user=request.user, movie=movie).exists():
+                Rating.objects.filter(user=request.user, movie=movie).delete()
+                Rating.objects.create(user=request.user, rating=rating, movie=movie)
+                messages.success(request, 'You have changed your rating!')
+            else:
+                Rating.objects.create(user=request.user, rating=rating, movie=movie)
+                messages.success(request, 'Thank you for your review of the movie!')
             return HttpResponseRedirect(reverse('detail', args=(slug,)))
     else:
         form = RatingForm()
